(PC-19096) feat(navigation): Create component for TabBar title
