+++ src/features/navigation/TabBar/TabBarComponent.tsx
@@ -1,16 +1,16 @@
-import React, { ElementType } from 'react'
+import React from 'react'
 import { Platform } from 'react-native'
-import styled, { useTheme } from 'styled-components/native'
+import styled from 'styled-components/native'
 
 import { menu } from 'features/navigation/TabBar/routes'
+import { TabBarTitle as Title } from 'features/navigation/TabBar/TabBarTitle'
 import { TabRouteName } from 'features/navigation/TabBar/types'
-import { HiddenAccessibleText } from 'ui/components/HiddenAccessibleText'
 import { InternalTouchableLink } from 'ui/components/touchableLink/InternalTouchableLink'
 import { InternalNavigationProps } from 'ui/components/touchableLink/types'
 import { BicolorLogo } from 'ui/svg/icons/BicolorLogo'
 import { BicolorSelector } from 'ui/svg/icons/BicolorSelector'
 import { BicolorIconInterface } from 'ui/svg/icons/types'
-import { Spacer, getSpacing, Typo } from 'ui/theme'
+import { Spacer, getSpacing } from 'ui/theme'
 
 const SELECTOR_WIDTH = '80%'
 const SELECTOR_HEIGHT = getSpacing(1)
@@ -31,33 +31,29 @@ export const TabBarComponent: React.FC<Props> = ({
   navigateTo,
   onPress,
   tabName,
-}) => {
-  const { tabBar } = useTheme()
-  const TitleComponent = (tabBar.showLabels ? Title : HiddenAccessibleText) as ElementType
-  return (
-    <TabComponentContainer
-      navigateTo={navigateTo}
-      onBeforeNavigate={onPress}
-      activeOpacity={1}
-      selected={isSelected}
-      accessibilityLabel={menu[tabName].accessibilityLabel}
-      testID={menu[tabName].accessibilityLabel || menu[tabName].displayName}
-      aria-current={isSelected ? 'page' : undefined}>
-      {!!isSelected && (
-        <BicolorSelector
-          width={SELECTOR_WIDTH}
-          height={SELECTOR_HEIGHT}
-          testID={`${menu[tabName].accessibilityLabel} sélectionné`}
-        />
-      )}
-      <Spacer.Flex />
-      <StyledIcon as={BicolorIcon} selected={isSelected} />
-      <TitleComponent selected={isSelected}>{menu[tabName].displayName}</TitleComponent>
-      <Spacer.Flex />
-      {!!isSelected && <BicolorSelectorPlaceholder />}
-    </TabComponentContainer>
-  )
-}
+}) => (
+  <TabComponentContainer
+    navigateTo={navigateTo}
+    onBeforeNavigate={onPress}
+    activeOpacity={1}
+    selected={isSelected}
+    accessibilityLabel={menu[tabName].accessibilityLabel}
+    testID={menu[tabName].accessibilityLabel || menu[tabName].displayName}
+    aria-current={isSelected ? 'page' : undefined}>
+    {!!isSelected && (
+      <BicolorSelector
+        width={SELECTOR_WIDTH}
+        height={SELECTOR_HEIGHT}
+        testID={`${menu[tabName].accessibilityLabel} sélectionné`}
+      />
+    )}
+    <Spacer.Flex />
+    <StyledIcon as={BicolorIcon} selected={isSelected} />
+    <Title selected={isSelected} displayName={menu[tabName].displayName} />
+    <Spacer.Flex />
+    {!!isSelected && <BicolorSelectorPlaceholder />}
+  </TabComponentContainer>
+)
 
 const StyledIcon = styled(BicolorLogo).attrs<{ selected?: boolean }>(({ theme, selected }) => ({
   color: selected ? undefined : theme.colors.greyDark,
@@ -76,10 +72,4 @@ const TabComponentContainer = styled(InternalTouchableLink).attrs(
   alignItems: 'center',
   height: theme.tabBar.height,
   flex: 1,
-}))
-
-const Title = styled(Typo.Caption)<{ selected?: boolean }>(({ theme, selected }) => ({
-  color: selected ? theme.colors.black : theme.colors.greyDark,
-  fontSize: theme.tabBar.fontSize,
-  textAlign: 'center',
 }))
